feat: rebase 이미지 아이콘 충돌 해결
