+++ src/components/commons/buttons/index.ts
@@ -1,4 +1,3 @@
-export * from './ArrowButton';
 export * from './BaseButton';
 export * from './CountButton';
 export * from './MoreButton';
